feat: Add CRUD functionality for students
This commit adds CRUD (Create, Read, Update, Delete) functionality for students. It includes the creation of the student model, routes for retrieving a list of students and a specific student by ID, creating a new student, updating a student, and deleting a student. The changes are implemented in the estudianteController.js file and the index.js file.
